Add cmd to output version (#14)
This builds on work that Carl Quinn did and opened in a different PR.
His approach was slightly different. The entire Version class was
inline in the Makefile and echoed into existence.

With the approach here, we introduce the ability to do c preprocessor
style substitution into the project via %.pony.in files. Each will
be processed with all the .in rules as seen in the Makefile.

Currently there is only 1 .in rule: "substitute in version".

This along with a couple of other Makefile changes results in:

- corral being rebuilt anytime the VERSION file changes
- corral being rebuilt if the `version=FOO` make option is used
- corral not needing to be rebuilt otherwise

Issues with this approach

- Each generated file needs to be added to .gitignore

Closes #5