Passthru exit codes from Runners to early exit corral with same exit code
If a Runner critical to the correct result of corral fails then it is
better to have all of corral exit immediately (passing through the exit
code). For example, if a git pull fails to execute, we could continue to
compile the project with the wrong code ending with an incorrect result.
Another common example when ponyc compile fails and you are using corral
in a larger make system, you want the make system to know the command
failed.