adds new "exec" command, mainly to allow use of lldb with corral.
The "run" command captures all input and output of the command being
run, making it difficult to be used with dynamic tools such as lldb when
debugging ponyc changes. Instead of altering how run works, I added an
"exec" command which just execve's the process and gives everything
over to the command being executed. This has allowed full access for
working with lldb.