Baby steps towards making CmdUpdate unit testable
As part of the work on #120, an error when using the update command,
I wanted to add tests to demonstrate the bug and then show that the
bug was fixed by my changes.

This is problematic because there are no real unit tests for any
commands.

This commit is the first step in providing better unit testability
of various parts of corral in general and commands specifically.

The only test added by this commit is that when using a corral.json
with empty deps, that no VCS operations will be executed by _Updater.

There are a number of changes that are brought in my this initial refactor.
As we move forward with making corral more testable, some of these changes
will probably fall away. I did what I considered the minimal thing to get
a harness for testing _Updater working. This harness should be enough to
test using the various test-data bundles that we have as well as testing
issue #120.

The approach I have taken does leave a lot of things in an inconsistent state.
For example, this commit introduces a new way of organizing unit tests that keeps
said tests near the code they are testing and allows for testing private implementations
like the _Update that is part of the update cmd.

These inconsistencies are temporary and as we move forward with adding more tests, will
start to fall away.