Don't update dependencies more than once
Prior to this commit, when update would run, if there were transitive dependencies, it might try to fetch/update them more than once potentially leading to errors.

This commit adds a regression unit test that shows the problem and a simple fix to the _Updater to handle. I use "simple" to mean "small" or "baby-steps". The internal design of _Updater needs to be examined in more depth in order to make this all better.

I'm not tackling that now. Before getting into design, we need to add more tests in general to see what affordances for insight and tracing that corral needs to provide then we can build the design around that.

I wanted to add tests to demonstrate the bug and then show that the
bug was fixed by my changes. This was problematic because there are no real unit tests for any
commands.

This PR includes a first step in providing better unit testability
of various parts of corral in general and commands specifically.

There are a number of testing-related changes that are brought this PR.
As we move forward with making corral more testable, some of these changes
will probably fall away. I did what I considered the minimal thing to get
a unit-test harness for testing _Updater working. 

The approach I have taken does leave a lot of things in an inconsistent state.
For example, this PR introduces a new way of organizing unit tests that keeps
said tests near the code they are testing and allows for testing private implementations
like the _Update that is part of the update cmd.

These inconsistencies are temporary and as we move forward with adding more tests, will
start to fall away.

Closes #120