Turn off cycle detector when running corral
Corral as currently constituted doesn't need the cycle detector so,
it ends up just slowly the app down. But not a slow down you would
notice.

The bigger issue is that there is a race condition in the runtime
at the moment that is present when the cycle detector is on and
running corral tests triggers it sometimes.

Until that issue is addressed, we should turn the cycle detector
off so we don't get failures from a known source that we don't need to
get them from.