Don't remove empty fields from configuration
Prior to this commit, optional fields were being removed from corral.json.
For things like "info" that we want people to fill out, this is detrimenal
as there is no template for what to fill out.

Given that this is only impacting on "info", I think it makes sense to
change the behavior.