Add new `packages` field to corral.json
The `packages` field is a new experimental field that will list all
packages that are part of a bundle that should be included for distribution.

At the moment, entries must be added by hand, this commit assures that
any that exist are not removed when corral manipulates the corral.json file.

Eventually, if this feature holds, we will have commands to add and remove
packages as well as a command to prepare for release. Releases, rather than
being a "repo over here with tag" can be hosted on services like GitHub packages
or cloudsmith and would only include the bare minumum needed for the package.

Initially, this field is being used to generate documentation that works across
library projects. Existing exploratory work for that is being done in the
ponylang/library-documentation-action repository.