Step towards turning corral into a library plus application
We want to be able to use corral functionality like extracting information
from corral.json from other applications without those applications rolling
their own parsing. The first step in that direction are these changes to
make corral look like a library that can be used as a dependency using
corral.

There's a lot of questions that will need to be answered as we move down this
path. For example, we've had "library Makefiles" and "application Makefiles".
We currently have "library documentation generation" that is the output of
`make docs` i.e. pony doc content; when a repo hosts an application and a library
bundle, the docs should really be not just the API docs but other docs as well.

How does all this get worked out? We'll see as we move forward.