Switch supported FreeBSD to 13.0
FreeBSD 13.0 has been released. This commit switches us from supporting
corral on FreeBSD 12.2 to 13.0.

- CI has been switched to run on FreeBSD 13.0.
- Releases are now built on FreeBSD 13.0

As part of this change, as there is no ponyc release for FreeBSD 13 yet,
we are using nightly ponyc builds for testing and building. Once the
next version of ponyc is released, the freebsd-13.0-install-pony-tools.bash
script should be updated from "nightlies" to "releases".