Add pack command
This commit adds a new pack command that will be allow users to create "corral archives" for distributing pony libraries rather than relying on keeping git repos up-to-date.

This commit adds the corral archive encoder and decoder for a beta version. I want to get this merged now so I can start doing testing with it.

There's no release notes or changelog yet as there are several other steps before this gets "turned on for the public".