Fix resolving relative paths with `corral run` (#216)
The logic to determine the absolute path to an executable was using `FilePath.from(base, subdir)` which errors if the resulting path is not below the `base`. This is the case if `subdir` is e.g. `../foo`.

This fix stems from trying to execute:

```
$ corral run -- ../../ponyc/build/debug/ponyc
run: couldn't run program: ../../ponyc/build/debug/ponyc .
```

The error output is clearly not really informative.
With this PR the above command actually runs the existing ponyc.

Other changes in here all related to improving `corral run`:
* First try to resolve relative paths against the current working directory, then against `$PATH`. This enables running relative paths to parent folders (See the above command).
* Display different errors if no args were provided or the executable to run could not be found