Fix bug with default bundle directory handling
There was a bug in corral's handling of the the `-d` command line
option. The default value was set to the string "<cwd>". This would
result in corral.json not being found.

The bug wasn't triggered because a bug in the Pony standard library
prevented the default value from correctly being applied. Once that
bug was fixed, corral stopped working unless a `-d` value was
provided.